TEIIDWEBTL-8 Fix for AddViewSources dialog dropdown not populating
