+++ teiid-webui-webapp/src/main/java/org/teiid/webui/client/messages/messages.json
@@ -44,6 +44,8 @@
   "createdataservice.serviceDescriptionTextBox.tooltip" : "A description for the Data Service",
   "createdataservice.createServiceButton.tooltip" : "Create the Data Service",
   "createdataservice.cancelButton.tooltip" : "Cancel creation of the Data Service",
+  "createdataservice.error-getting-svcsources" : "Error fetching the Data Sources",
+  "editdataservice.error-getting-svcsources" : "Error fetching the Data Sources",
   "editdataservice.error-getting-svc-details" : "Error retrieving the service details",
   "editdataservice.status-label-enter-name" : "Please enter a name for your service",
   "editdataservice.status-label-click-save" : "Click 'Save Changes' to save your service",
@@ -173,7 +175,6 @@
   "ssource-editor-panel.dsNamesTable.tooltip" : "All sources available for use in your services",
   "ssource-editor-panel.tablesAndProcsTable.tooltip" : "All tables and procedures in the selected source",
   "ssource-editor-panel.columnsTable.tooltip" : "All columns in the selected table or procedure",
-  "ssource-editor-panel.error-getting-svcsources" : "Error populating the Data Sources",
   "ssource-editor-panel.error-getting-tables-procs" : "Error populating the tables and procedures",
   "ssource-editor-panel.error-getting-tablecols" : "Error populating the columns",
   "joineditor.lhs-table.title" : "LHS Table",
