+++ teiid-webui-webapp/src/main/java/org/teiid/webui/client/screens/CreateDataServiceScreen.html
@@ -44,8 +44,8 @@
             <input type="text" style="width:100%" data-field="textbox-create-service-name" />
           </div>
           <div class="col-sm-3">
-            <label title="Tags for identification of the Data Service"><strong>Tags</strong></label>
-            <input type="text" style="width:100%" placeholder="Not yet implemented" />
+            <div style="height:20px"></div>
+            <label class="checkbox"><input type="checkbox" data-field="checkbox-create-service-visibility"> Service visible to outside parties</label>
           </div>
           <div class="col-sm-6"></div>
         </div>
@@ -55,11 +55,7 @@
             <label title="A description for the Data Service"><strong>Description</strong></label>
             <textarea data-field="textarea-create-service-description" rows="2" style="width:100%" ></textarea>
           </div>
-          <div class="col-sm-3">
-            <div style="height:20px"></div>
-            <label class="checkbox"><input type="checkbox" data-field="checkbox-create-service-visibility"> Service visible to outside parties</label>
-          </div>
-          <div class="col-sm-6"></div>
+          <div class="col-sm-9"></div>
         </div>
         
         <div class="row">
