+++ teiid-webui-webapp/src/main/java/org/teiid/webui/client/screens/EditDataServiceScreen.html
@@ -44,8 +44,8 @@
             <input type="text" data-field="textbox-edit-service-name" style="width:100%" />
           </div>
           <div class="col-sm-3">
-            <label title="Tags for identification of the Data Service"><strong>Tags</strong></label>
-            <input type="text" placeholder="Not yet implemented" style="width:100%" />
+            <div style="height:20px"></div>
+            <label class="checkbox"><input type="checkbox" data-field="checkbox-edit-service-visibility"> Service visible to outside parties</label>
           </div>
           <span class="col-sm-6"></span>
         </div>
@@ -55,11 +55,7 @@
             <label title="A description for the Data Service"><strong>Description</strong></label>
             <textarea data-field="textarea-edit-service-description" rows="2" style="width:100%" ></textarea>
           </div>
-          <div class="col-sm-3">
-            <div style="height:20px"></div>
-            <label class="checkbox"><input type="checkbox" data-field="checkbox-edit-service-visibility"> Service visible to outside parties</label>
-          </div>
-          <span class="col-sm-6"></span>
+          <span class="col-sm-9"></span>
         </div>
                 
         <div class="row">
